berufsschule-gateway/requirements.txt
asyncpg~=0.29.0
pip~=24.0
python-dotenv~=1.0.1